# storage: reduce allocations when encoding MVCCValue's 