--- a/pkg/storage/bench_test.go
+++ b/pkg/storage/bench_test.go
@@ -218,6 +218,25 @@ func BenchmarkMVCCExportToSST(b *testing.B) {
 			runMVCCExportToSST(b, opts)
 		})
 	}
+	withImportEpochs := []bool{false, true}
+	for _, ie := range withImportEpochs {
+		numKey := numKeys[len(numKeys)-1]
+		numRevision := numRevisions[len(numRevisions)-1]
+		numRangeKey := numRangeKeys[len(numRangeKeys)-1]
+		exportAllRevisionVal := exportAllRevisions[len(exportAllRevisions)-1]
+		b.Run(fmt.Sprintf("importEpochs=%t/numKeys=%d/numRevisions=%d/exportAllRevisions=%t",
+			ie, numKey, numRevision, exportAllRevisionVal,
+		), func(b *testing.B) {
+			opts := mvccExportToSSTOpts{
+				numKeys:            numKey,
+				numRevisions:       numRevision,
+				numRangeKeys:       numRangeKey,
+				exportAllRevisions: exportAllRevisionVal,
+				importEpochs:       ie,
+			}
+			runMVCCExportToSST(b, opts)
+		})
+	}
 }
 
 const numIntentKeys = 1000
@@ -1678,8 +1697,8 @@ func runMVCCAcquireLockCommon(
 }
 
 type mvccExportToSSTOpts struct {
-	numKeys, numRevisions, numRangeKeys     int
-	exportAllRevisions, useElasticCPUHandle bool
+	numKeys, numRevisions, numRangeKeys                   int
+	importEpochs, exportAllRevisions, useElasticCPUHandle bool
 
 	// percentage specifies the share of the dataset to export. 100 will be a full
 	// export, disabling the TBI optimization. <100 will be an incremental export
@@ -1741,6 +1760,9 @@ func runMVCCExportToSST(b *testing.B, opts mvccExportToSSTOpts) {
 		for j := 0; j < opts.numRevisions; j++ {
 			mvccKey := MVCCKey{Key: key, Timestamp: hlc.Timestamp{WallTime: mkWall(j), Logical: 0}}
 			mvccValue := MVCCValue{Value: roachpb.MakeValueFromString("foobar")}
+			if opts.importEpochs {
+				mvccValue.ImportEpoch = 1
+			}
 			err := batch.PutMVCC(mvccKey, mvccValue)
 			if err != nil {
 				b.Fatal(err)
@@ -1792,14 +1814,15 @@ func runMVCCExportToSST(b *testing.B, opts mvccExportToSSTOpts) {
 		startTS := hlc.Timestamp{WallTime: startWall}
 		endTS := hlc.Timestamp{WallTime: endWall}
 		_, _, err := MVCCExportToSST(ctx, st, engine, MVCCExportOptions{
-			StartKey:           MVCCKey{Key: keys.LocalMax},
-			EndKey:             roachpb.KeyMax,
-			StartTS:            startTS,
-			EndTS:              endTS,
-			ExportAllRevisions: opts.exportAllRevisions,
-			TargetSize:         0,
-			MaxSize:            0,
-			StopMidKey:         false,
+			StartKey:               MVCCKey{Key: keys.LocalMax},
+			EndKey:                 roachpb.KeyMax,
+			StartTS:                startTS,
+			EndTS:                  endTS,
+			ExportAllRevisions:     opts.exportAllRevisions,
+			TargetSize:             0,
+			MaxSize:                0,
+			StopMidKey:             false,
+			IncludeMVCCValueHeader: opts.importEpochs,
 		}, &buf)
 		if err != nil {
 			b.Fatal(err)